Dynamic evaluation: Get decision from DNA by name when `require_hyper_name` is True.
This allows the user to use one function for defining the search space, and another function for applying the decisions, given the name matches. This CL is desirable due to the fact that the function executed by `pg.hyper.trace` usually need to be lighter than the one that actual runs. With this CL, we don't need to worry about the declaration/running order of the decision points.

PiperOrigin-RevId: 463193822