Introducing ext/mutfun.
`pg.ext.mutfun` is introduced to make it easy to search for an arbitrary function with pre-defined instructions. This first check-in enables the interface for symbolic programs based on `Instruction` and `Function`, together with basic math operations. We may add more operations as well as control flow operations such as `If/For`.

PiperOrigin-RevId: 515095344